refactor(plugins): adopt folder-per-plugin layout (closes #36)
Each ingestion source plugin and enrichment provider now lives in its
own folder containing the implementation file, a one-line `__init__.py`
that re-exports the module for the existing pkgutil-based discovery, a
plugin-specific README.md documenting the content type / requirements /
config schema, and a co-located test_<name>.py.

- Ten source plugins migrated under src/ingestion/sources/<name>/
- Three enrichment providers migrated under src/enrichment/providers/<name>/
- _rom_title.py helper relocated alongside the roms plugin that uses it
- arr_base.py kept flat as the shared base class for Radarr and Sonarr
- pytest testpaths extended to include src/, mypy excludes test_*.py
- ARCHITECTURE.md, CLAUDE.md, CONTRIBUTING.md, and PLUGIN_DEVELOPMENT.md
  updated to describe the new layout
- Test patches updated from package-path to deeper module-path so they
  patch the correct binding of `requests` etc. inside each plugin module
