+++ CLAUDE.md
@@ -33,14 +33,15 @@ src/
 ├── cli/              # Click CLI interface
 ├── web/              # FastAPI web interface
 │   └── static/
-│       ├── themes/     # UI themes (folder-per-theme, auto-discovered)
+│       ├── themes/     # UI themes (folder-per-theme, each with README + theme.json + colors.css)
 │       └── dist/       # Built Vue SPA (generated by Vite, gitignored)
 ├── ingestion/        # Data ingestion
-│   └── sources/      # Source plugins (auto-discovered)
+│   └── sources/      # Source plugins (folder-per-plugin: <name>/<name>.py + README.md + test_<name>.py)
 ├── llm/              # Ollama interaction
 ├── storage/          # SQLite + ChromaDB
 ├── recommendations/  # Recommendation engine (scorers, pipeline, ranking, genre_clusters)
 ├── enrichment/       # Background metadata enrichment
+│   └── providers/    # Enrichment providers (folder-per-provider, same layout as sources)
 ├── conversation/     # Conversational AI chat system
 ├── models/           # Data models (ContentItem, ContentType, UserPreferenceConfig)
 └── utils/            # Utility functions (list_merge, series, sorting)
@@ -49,7 +50,8 @@ resources/            # Frontend source (Vue 3 + Tailwind CSS v4)
 └── css/              # CSS variables (base.css) and Tailwind config (tailwind.css)
 index.html            # Vite SPA entry point
 vite.config.ts        # Vite build configuration
-tests/                # Mirrors src/ structure
+tests/                # Cross-cutting tests (CLI, web, storage, recommendations, conversation).
+                      # Plugin-local tests live next to the plugin: src/.../<plugin>/test_<plugin>.py.
 config/               # Configuration files (example.yaml for tests)
 docker/               # Container helpers
 ├── entrypoint.sh         # First-run config.yaml bootstrap (recommendinator image)
