+++ src/enrichment/providers/openlibrary/__init__.py
@@ -1,0 +1,3 @@
+"""openlibrary enrichment provider package."""
+
+from src.enrichment.providers.openlibrary.openlibrary import *  # noqa: F401, F403
