+++ src/enrichment/providers/rawg/__init__.py
@@ -1,0 +1,3 @@
+"""rawg enrichment provider package."""
+
+from src.enrichment.providers.rawg.rawg import *  # noqa: F401, F403
