+++ src/enrichment/providers/tmdb/__init__.py
@@ -1,0 +1,3 @@
+"""tmdb enrichment provider package."""
+
+from src.enrichment.providers.tmdb.tmdb import *  # noqa: F401, F403
