+++ src/ingestion/sources/epic_games/__init__.py
@@ -1,0 +1,3 @@
+"""epic_games plugin package."""
+
+from src.ingestion.sources.epic_games.epic_games import *  # noqa: F401, F403
