+++ src/ingestion/sources/generic_csv/__init__.py
@@ -1,0 +1,3 @@
+"""generic_csv plugin package."""
+
+from src.ingestion.sources.generic_csv.generic_csv import *  # noqa: F401, F403
