+++ src/ingestion/sources/generic_json/__init__.py
@@ -1,0 +1,3 @@
+"""generic_json plugin package."""
+
+from src.ingestion.sources.generic_json.generic_json import *  # noqa: F401, F403
