+++ src/ingestion/sources/gog/__init__.py
@@ -1,0 +1,3 @@
+"""gog plugin package."""
+
+from src.ingestion.sources.gog.gog import *  # noqa: F401, F403
