+++ src/ingestion/sources/goodreads/__init__.py
@@ -1,0 +1,3 @@
+"""goodreads plugin package."""
+
+from src.ingestion.sources.goodreads.goodreads import *  # noqa: F401, F403
