+++ src/ingestion/sources/markdown/__init__.py
@@ -1,0 +1,3 @@
+"""markdown plugin package."""
+
+from src.ingestion.sources.markdown.markdown import *  # noqa: F401, F403
