+++ src/ingestion/sources/radarr/__init__.py
@@ -1,0 +1,3 @@
+"""radarr plugin package."""
+
+from src.ingestion.sources.radarr.radarr import *  # noqa: F401, F403
