+++ src/ingestion/sources/roms/__init__.py
@@ -1,0 +1,3 @@
+"""roms plugin package."""
+
+from src.ingestion.sources.roms.roms import *  # noqa: F401, F403
