+++ src/ingestion/sources/roms/test_rom_title.py
@@ -6,7 +6,7 @@ import re
 
 import pytest
 
-from src.ingestion.sources._rom_title import (
+from src.ingestion.sources.roms._rom_title import (
     clean_display_title,
     compile_extra_patterns,
     normalize_title_key,
