+++ src/ingestion/sources/sonarr/__init__.py
@@ -1,0 +1,3 @@
+"""sonarr plugin package."""
+
+from src.ingestion.sources.sonarr.sonarr import *  # noqa: F401, F403
