+++ src/ingestion/sources/sonarr/test_sonarr.py
@@ -6,7 +6,7 @@ import pytest
 import requests
 
 from src.ingestion.plugin_base import SourceError, SourcePlugin
-from src.ingestion.sources.sonarr import SonarrPlugin
+from src.ingestion.sources.sonarr.sonarr import SonarrPlugin
 from src.models.content import ConsumptionStatus, ContentType
