+++ src/ingestion/sources/steam/__init__.py
@@ -1,0 +1,3 @@
+"""steam plugin package."""
+
+from src.ingestion.sources.steam.steam import *  # noqa: F401, F403
